requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.12.0  # CPU version is better for i3 to avoid GPU overhead errors
mediapipe==0.10.8
scikit-learn==1.3.0
pandas==2.0.3
opencv-python-headless
transformers